# api/resources/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Acesso ao sistema (US-01)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Acesso concedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas"
          },
          "403": {
            "description": "Acesso bloqueado temporariamente"
          }
        },
        "tags": [
          "Autenticação"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "summary": "Renovar sessão a partir do refresh token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Novos tokens emitidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Sessão expirada ou inválida"
          }
        },
        "tags": [
          "Autenticação"
        ]
      }
    },
...
    "/attempts/available": {
      "get": {
        "operationId": "AttemptsController_available",
        "summary": "Avaliações disponíveis para mim (US-13)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AvailableAssessmentDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Tentativas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/attempts/my": {
      "get": {
        "operationId": "AttemptsController_my",
        "summary": "Minhas avaliações e situação (US-29)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttemptStatusDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Tentativas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/attempts": {
      "post": {
        "operationId": "AttemptsController_start",
        "summary": "Iniciar avaliação (US-14)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartAttemptDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Processo indisponível ou fora da janela"
          },
          "403": {
            "description": "Não vinculado ao processo"
          }
        },
        "tags": [
          "Tentativas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/attempts/{id}": {
      "get": {
        "operationId": "AttemptsController_detail",
        "summary": "Responder avaliação — questões e tempo restante (US-14/15/18)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptDetailDto"
                }
              }
            }
          },
          "403": {
            "description": "Avaliação de outro candidato"
          }
        },
        "tags": [
          "Tentativas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/attempts/{id}/submit": {
      "patch": {
        "operationId": "AttemptsController_submit",
        "summary": "Finalizar e enviar avaliação (US-19)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptStatusDto"
                }
              }
            }
          },
          "400": {
            "description": "Avaliação já finalizada"
          }
        },
        "tags": [
          "Tentativas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/attempts/{id}/tab-exit": {
      "post": {
        "operationId": "AttemptsController_tabExit",
        "summary": "Registrar saída de janela durante a prova (US-21)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Saída registrada"
          }
        },
        "tags": [
          "Tentativas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/attempts/{id}/result": {
      "get": {
        "operationId": "AttemptsController_result",
        "summary": "Resultado detalhado da minha avaliação (US-30)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhe da avaliação corrigida"
          }
        },
        "tags": [
          "Tentativas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/attempts/{id}/extra-time": {
      "patch": {
        "operationId": "AttemptsController_extraTime",
        "summary": "Conceder tempo extra ao candidato (US-12)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtraTimeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tempo atualizado"
          },
          "400": {
            "description": "Avaliação já finalizada"
          }
        },
        "tags": [
          "Tentativas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/attempts/{id}/approve": {
      "patch": {
        "operationId": "AttemptsController_approve",
        "summary": "Aprovar/reprovar manualmente (US-27)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Decisão registrada"
          }
        },
        "tags": [
          "Tentativas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/attempts/{id}/feedback": {
      "patch": {
        "operationId": "AttemptsController_feedback",
        "summary": "Registrar feedback geral (US-28)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Feedback registrado"
          }
        },
        "tags": [
          "Tentativas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/attempts/{id}/correction": {
      "get": {
        "operationId": "AttemptsController_correction",
        "summary": "Tentativa para correção — respostas e questões abertas",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados de correção"
          }
        },
        "tags": [
          "Tentativas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/answers": {
      "put": {
        "operationId": "AnswersController_save",
        "summary": "Salvamento automático de resposta (US-16)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveAnswerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resposta salva"
          },
          "400": {
            "description": "Avaliação finalizada ou tempo encerrado"
          },
          "403": {
            "description": "Avaliação de outro candidato"
          }
        },
        "tags": [
          "Respostas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/answers/{id}/grade": {
      "put": {
        "operationId": "AnswersController_grade",
        "summary": "Corrigir questão aberta (US-24)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GradeAnswerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Correção registrada"
          },
          "400": {
            "description": "Pontuação acima do peso ou questão não aberta"
          },
          "404": {
            "description": "Resposta não encontrada"
          }
        },
        "tags": [
          "Respostas"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "API — Software de Processo Avaliativo",
    "description": "API REST para condução de processos avaliativos (avaliador e avaliado). Autenticação via JWT no cabeçalho Authorization: Bearer <token>.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@avaliacao.local"
          },
          "senha": {
            "type": "string",
            "example": "Admin@123"
          }
        },
        "required": [
          "email",
          "senha"
        ]
      },
      "UsuarioResumoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "perfil": {
            "type": "string",
            "enum": [
              "ADMIN",
              "AVALIADOR",
              "AVALIADO"
            ]
          }
        },
        "required": [
          "id",
          "nome",
          "email",
          "perfil"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Token de acesso JWT"
          },
          "refreshToken": {
            "type": "string",
            "description": "Token de atualização JWT"
          },
          "usuario": {
            "$ref": "#/components/schemas/UsuarioResumoDto"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "usuario"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Token de atualização emitido no login"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "AvailableAssessmentDto": {
        "type": "object",
        "properties": {
          "processoId": {
            "type": "string",
            "format": "uuid"
          },
          "titulo": {
            "type": "string"
          },
          "descricao": {
            "type": "object",
            "nullable": true
          },
          "tempoMinutos": {
            "type": "number"
          },
          "totalPerguntas": {
            "type": "number"
          }
        },
        "required": [
          "processoId",
          "titulo",
          "tempoMinutos",
          "totalPerguntas"
        ]
      },
      "AttemptStatusDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "processoTitulo": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "EM_ANDAMENTO",
              "ENVIADA",
              "CORRIGIDA"
            ]
          },
          "situacao": {
            "type": "string",
            "description": "Situação visível ao candidato"
          },
          "scoreParcial": {
            "type": "object",
            "nullable": true
          },
          "scoreTotal": {
            "type": "object",
            "nullable": true
          },
          "aprovado": {
            "type": "object",
            "nullable": true
          },
          "abertasPendentes": {
            "type": "number"
          },
          "finalizadoEm": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "processoTitulo",
          "status",
          "situacao",
          "abertasPendentes"
        ]
      },
      "StartAttemptDto": {
        "type": "object",
        "properties": {
          "processoId": {
            "type": "string",
            "format": "uuid",
            "description": "Processo a ser respondido"
          }
        },
        "required": [
          "processoId"
        ]
      },
      "QuestaoTentativaDto": {
        "type": "object",
        "properties": {
          "perguntaId": {
            "type": "string",
            "format": "uuid"
          },
          "enunciado": {
            "type": "string"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "ABERTA",
              "FECHADA"
            ]
          },
          "peso": {
            "type": "number"
          },
          "instrucoes": {
            "type": "object",
            "nullable": true
          },
          "ordem": {
            "type": "number"
          },
          "alternativas": {
            "type": "array",
            "description": "Alternativas (sem indicar a correta) para perguntas fechadas",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "texto": {
                  "type": "string"
                }
              }
            }
          },
          "respostaAtual": {
            "type": "object",
            "description": "Resposta já salva pelo candidato"
          }
        },
        "required": [
          "perguntaId",
          "enunciado",
          "tipo",
          "peso",
          "ordem"
        ]
      },
      "AttemptDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "EM_ANDAMENTO",
              "ENVIADA",
              "CORRIGIDA"
            ]
          },
          "iniciadoEm": {
            "format": "date-time",
            "type": "string"
          },
          "tempoRestanteSegundos": {
            "type": "number",
            "description": "Segundos restantes para o fim da prova"
          },
          "questoes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestaoTentativaDto"
            }
          }
        },
        "required": [
          "id",
          "status",
          "iniciadoEm",
          "tempoRestanteSegundos",
          "questoes"
        ]
      },
      "ExtraTimeDto": {
        "type": "object",
        "properties": {
          "minutos": {
            "type": "number",
            "example": 10,
            "description": "Minutos extras a conceder"
          }
        },
        "required": [
          "minutos"
        ]
      },
      "ApproveDto": {
        "type": "object",
        "properties": {
          "aprovado": {
            "type": "boolean",
            "example": true,
            "description": "true = aprovado, false = reprovado"
          }
        },
        "required": [
          "aprovado"
        ]
      },
      "FeedbackDto": {
        "type": "object",
        "properties": {
          "feedbackGeral": {
            "type": "string",
            "example": "Desempenho consistente nas questões de API."
          }
        },
        "required": [
          "feedbackGeral"
        ]
      },
...
      "SaveAnswerDto": {
        "type": "object",
        "properties": {
          "tentativaId": {
            "type": "string",
            "format": "uuid"
          },
          "perguntaId": {
            "type": "string",
            "format": "uuid"
          },
          "textoResposta": {
            "type": "string",
            "description": "Resposta de pergunta aberta"
          },
          "alternativaId": {
            "type": "string",
            "format": "uuid",
            "description": "Alternativa escolhida (pergunta fechada)"
          }
        },
        "required": [
          "tentativaId",
          "perguntaId"
        ]
      },
      "GradeAnswerDto": {
        "type": "object",
        "properties": {
          "pontos": {
            "type": "number",
            "example": 8,
            "description": "Pontos atribuídos (0 até o peso da questão)"
          },
          "observacao": {
            "type": "string",
            "example": "Boa resposta, faltou citar exemplos."
          }
        },
        "required": [
          "pontos"
        ]
      }
    }
  }
}